Refactor Pyodide worker runtime
